requirements.txt
# Core
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-WTF==1.2.2

# Database
SQLAlchemy==2.0.36

# Server
gunicorn==23.0.0

# Utilities
python-dotenv==1.0.1
Werkzeug==3.1.3

# Production extras (descomentar en VPS)
# psycopg2-binary==2.9.10
# redis==5.2.1
# celery==5.4.0
